requirements.txt
streamlit==1.50.0
python-dotenv==1.2.1

langchain==0.3.27
langchain-community==0.3.31
langchain-openai==0.3.35

openai==2.24.0
faiss-cpu==1.13.0
tiktoken==0.12.0